requirements.txt
flask
python-dotenv
supabase
requests
gunicorn
